Online Food Order System
This Online Restaurant Management System is in PHP, JavaScript, and CSS. Talking about the features of this system, it contains the admin section and the user (customer) section. All the editing’s, updating, managing order details, food items, and categories can be managed from the admin side while users can go through the homepage, about, and contact pages. In order to order the food items, the user has to create an account and sign in or log in.

Front end: HTML, CSS, JavaScript, PHP
Backend: Microsoft SQL
Run project on WAMP